# TuOpenBank/v1/swagger-v1.yaml
openapi: 3.0.0
info:
  title: Techu Open API Bank
  description: Open API Bank
  version: '1.0.0'
# Added by API Auto Mocking Plugin
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Pabloin/TuOpenBank/1.0.0
paths:
  /user:
    post:
      x-swagger-router-controller: User
      operationId: createUser
      tags:
        - /user
      summary: Crea un usuario
      description: Alta de un usuario con credenciales de login validas
      requestBody:
        required: true
        description: Objecto JSON que contiene un usuario
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          description: Operación exitosa
  /user/login:
    get:
      x-swagger-router-controller: User
      operationId: loginUser
      tags:
        - /user
      summary: Login del usuario en el sistema
      description: ""
      parameters:
      - name: username
        in: query
        description: Nombre del usuario para el login
        required: true
        schema:
          type: string
      - name: password
        in: query
        description: Password del usuario sin encriptar
        required: true
        schema:
          type: string
      responses:
        '200':
          description: operacion exitosa
          content:
            application/json:
              schema:
                type: string
          headers:
            X-Rate-Limit:
              schema:
                type: integer
                format: int32
              description: Invocaciones permitidas por el usuario por hora
            X-Expires-After:
              schema:
                type: string
                format: date-time
              description: date in UTC cuando el token
  /user/logout:
    get:
      x-swagger-router-controller: User
      operationId: logoutUser
      tags:
        - /user
      summary: Logout del usuario del sistema
      description: ""
      parameters: []
      responses:
        default:
          description: Logout exitoso.
  /account/{accountId}:
    get:
      x-swagger-router-controller: Account
      operationId: getAccount
      tags:
        - /account
      summary: Account detail
      description: Retorna la informacion de una cuenta
      parameters:
        - name: accountId
          in: path
          description: Id de la cuenta
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: retorna info de una cuenta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
  /product/{productId}:
    get:
      x-swagger-router-controller: Product
      operationId: getProducts
      tags:
        - /product
      summary: detalle de un producto
      description: Retorna la informacion de un producto
      parameters:
        - name: productId
          in: path
          description: Id del producto
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: retorna info de un producto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /quote/{quoteId}:
    get:
      x-swagger-router-controller: Quote
      operationId: getQuote
      tags:
        - /quote
      summary: Mensajes, textos o promociones
      description: Retorna un mensaje random para el usuario
      parameters:
        - name: quoteId
          in: path
          description: QuoteId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: retorna un mensaje
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
  /quote/protected/{quoteId}:
    get:
      x-swagger-router-controller: Quote
      operationId: getProtectedQuote
      tags:
        - /quote
      summary: Mensajes, textos o promociones
      description: Retorna un mensaje random para el usuario
      parameters:
        - name: quoteId
          in: path
          description: QuoteId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: retorna un mensaje
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
  /transaction/{userId}:
    get:
      x-swagger-router-controller: Transaction
      operationId: getTransaction
      tags:
        - /transaction
      summary: lista de transacciones de un usuario
      description: Retorna la lista de transacciones de una usuario
      parameters:
        - name: userId
          in: path
          description: UserId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: retorna lista de transacciones de un usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionList'
components:
  schemas:
    User:
      properties:
        userId:
          type: integer
        username:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        isLogged:
          type: boolean
          default: false
    Account:
      properties:
        userId:
          type: integer
        accountId:
          type: integer
        accountType:
          type: string
        accountBranch:
          type: string
        accountNumber:
          type: string
        accountDV:
          type: string
        accountCurrency:
          type: string
        accountBalance:
          type: integer
    Product:
      properties:
        productId:
          type: integer
        productCode:
          type: string
        productDescription:
          type: string
        productLabel:
          type: string
        productCurrency:
          type: string
        productAmount:
          type: integer
    Transaction:
      properties:
        transactionId:
          type: integer
        userId:
          type: integer
        accountId:
          type: integer
        transactionCurrency:
          type: string
        transactionBalance:
          type: integer
        timestamp:
          type: string
    TransactionList:
      type: object
      properties:
        products:
          description: Transaction List
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    Quote:
      properties:
        quoteId:
          get:
          type: integer
        text:
          type: string